feat: return unsigned long (uint64)

+++ D02GiftShop/GiftShop.cs
@@ -3,16 +3,16 @@ namespace D02GiftShop;
 
 public class GiftShop
 {
-    public static int AddInvalidIds(IEnumerable<string> ranges)
+    public static ulong AddInvalidIds(IEnumerable<string> ranges)
     {
-        var sum = 0;
+        var sum = 0UL;
 
         foreach (var range in ranges)
         {
             var rangeLimits = range.Split('-');
 
-            var left = int.Parse(rangeLimits[0]);
-            var right = int.Parse(rangeLimits[1]);
+            var left = ulong.Parse(rangeLimits[0]);
+            var right = ulong.Parse(rangeLimits[1]);
 
             for (var number = left; number <= right; number++)
             {
